add api logic
