+++ .env
@@ -1,3 +1,4 @@
 VITE_APP_API_URL=http://localhost:3000
 VITE_API_URL=https://api.gala.fundacja.hematologiczna.org/api
+PAYMENT_API_URL= https://payments.fundacja.hematologiczna.org/api
 VITE_HCAPTCHA_SITE_KEY=dfd4a7f6-1c36-4914-b44c-0b9220d2059a
