+++ src/api/services.js
@@ -2,6 +2,7 @@ import axios from 'axios';
 import { client } from './fetchClient.js';
 
 const URL = import.meta.env.VITE_API_URL;
+const PAYMENT_API_URL = import.meta.env.PAYMENT_API_URL;
 
 //const URL = 'http://localhost:5000/api';
 
@@ -33,7 +34,6 @@ export const getImages = async () => {
   const response = await axios.get(URL + '/photos');
 
   // const response = await axios.get('http://localhost:5000/api/admin/users');
-  console.log(response.data);
   return response.data;
 };
 
@@ -44,16 +44,9 @@ export const getLogos = async () => {
   return data;
 };
 
-// export const checkPaymentStatus = async (paymentId) => {
-//   const data = await axios.get(
-//     'https://fundacja.hematologiczna.org/api/supporters',
-//   );
-//   return data;
-// };
-
 export const checkPaymentStatus = async (transactionId) => {
   const data = await axios.get(
-    `https://payments.fundacja.hematologiczna.org/api/payment/status/${transactionId}`,
+    `${PAYMENT_API_URL}/payment/status/${transactionId}`,
   );
   return data;
 };
@@ -83,6 +76,25 @@ export const addUser = async (userData) => {
         Accept: 'application/json',
       },
     });
+
+    return response.data;
+  } catch (error) {
+    console.error('Błąd płatności:');
+  }
+};
+
+export const createPayment = (payload) => {
+  try {
+    const response = axios.post(
+      `${PAYMENT_API_URL}/payment/createPayment`,
+      payload,
+      {
+        headers: {
+          'Content-Type': 'application/json',
+          Accept: 'application/json',
+        },
+      },
+    );
 
     return response.data;
   } catch (error) {
