--- src/pages/Registry/P24Form/p24Form.scss
@@ -1,146 +1,0 @@
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    
-   
-}
-
-b {
-    font-weight: bold;
-}
-
-.flex-wrapper {
-    display: flex;
-    justify-content: space-between;
-    flex-grow: 1;
-}
-  
-.P24Form {
-    border-radius: 20px;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    width: 100%;
-    max-width: 90%;
-   display: flex;
-   font-family: Poppins;
-  
-    flex-direction: column;
-    background-color: white;
-    background-color: rgb(245, 241, 241);
-
-    &-container {
-
-        width: 90%;
-        margin: 0 auto;
-        padding: 0;
-    }
-
-    &-title {
-        text-transform: uppercase;
-    }
-
-    &-input {
-        border-radius: 2px;
-    }
-    
-
-   
-
-    &__group {
-
-        &-title {
-           // padding: 15px 0;
-           font-weight: 500;
-           font-size: 24px;
-        }
-       
-
-      
-        &--1 {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-           // padding:20px 0 10px;
-        }
-
-        &--3 {
-          //  border: 1px solid red;
-            display: flex;
-            flex-direction: column;
-            gap:5px;
-            justify-content: space-between;
-            
-        }
-    }
-
-    &__buttons {
-        display: flex;
-        gap: 20px;
-        justify-content: center;
-    }
-
-    &-bottomLabel {
-        text-align: center;
-        display: flex;
-        flex-direction: column;
-        gap: 5px;
-    }
-
-    &-bottomMargin {
-        height: 50px;
-    }
-
-    & section {
-        padding: 16px 0;
-        
-        &:nth-child(n+2):nth-child(odd) {     
-            background-color: rgb(245, 241, 241);
-            display: grid;    
-        }
-
-        &:nth-child(even) {     
-            background-color: white;
-        }
-    }
-       
-
-    &-title {
-        margin-right: 20px;
-    }
-
-    &-logo {
-        width: 150px;
-        height: auto;
-    }
-
-   
-
-      
-}
-
-.wrapper {
-    display: flex;
-}
-  
-
-
-
-
-
-
-
-
-.desc {
-    margin-right:20px ;
-    padding: 0;
-}
-
-.border {
-    border: 1px solid red;
-}
