+++ src/pages/Registry/Registry.jsx
@@ -1,135 +1,22 @@
 import '../../styles/index.scss';
 import './registry.scss';
-import HCaptcha from '@hcaptcha/react-hcaptcha';
-import { useState, useRef, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { NavLink } from 'react-router-dom';
-import Checkbox from '../../components/Checkbox/Checkbox';
 import { useTranslation } from 'react-i18next';
-import { addUser, checkPaymentStatus } from '../../api/services';
+import { checkPaymentStatus } from '../../api/services';
 import Modal from './Modal/Modal.jsx';
-import P24Form from './P24Form/P24Form.jsx';
+import { RegistryDonateForm } from './RegistryDonateForm/RegistryDonateForm.jsx';
+
+import { RegistryForm } from './RegistryForm/RegistryForm.jsx';
 
 const Registry = () => {
-  const { i18n, t } = useTranslation();
-  const [donate, setDonate] = useState(100);
+  const { t } = useTranslation();
   const [searchParams] = useSearchParams();
-  const [isVerified, setIsVerified] = useState(false);
-  const [captchaSize, setCaptchaSize] = useState('normal');
-  const [p24formIsOpen, setP24FormIsOpen] = useState(false);
   const [modal, setModal] = useState({
     isOpen: false,
     event: '',
     url: null,
   });
-
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    street: '',
-    buildingNumber: '',
-    zipCode: '',
-    city: '',
-    lang: i18n.language,
-    consents: {
-      privacyPolicyAndRegulaminAccepted: false,
-      dataProcessingConsentAccepted: false,
-      marketingConsentAccepted: false,
-    },
-  });
-
-  const hCaptchaSiteKey = import.meta.env.VITE_HCAPTCHA_SITE_KEY;
-  const p24Images = Array.from({ length: 25 }, (_, i) => i + 1);
-
-  const handleConsentChange = (e) => {
-    const { name, checked } = e.target;
-
-    setFormData((prevData) => ({
-      ...prevData,
-      consents: {
-        ...prevData.consents,
-        [name]: checked,
-      },
-    }));
-  };
-
-  const handleHCaptcha = () => {
-    const res = hcaptchaRef.current.getResponse();
-
-    if (res) {
-      setIsVerified(true);
-    }
-  };
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prevState) => ({
-      ...prevState,
-      [name]: value,
-    }));
-  };
-
-  const hcaptchaRef = useRef(null);
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!isVerified) {
-      return alert('Proszę potwierdzić, że jesteś człowiekiem.');
-    }
-
-    try {
-      const response = await addUser(formData);
-
-      if (response && response.status === 'success') {
-        setModal({
-          isOpen: true,
-          event: 'registration-success',
-          url: response.paymentUrl,
-        });
-        setFormData({
-          name: '',
-          email: '',
-          street: '',
-          buildingNumber: '',
-          zipCode: '',
-          city: '',
-          lang: i18n.language,
-          consents: {
-            privacyPolicyAndRegulaminAccepted: false,
-            dataProcessingConsentAccepted: false,
-            marketingConsentAccepted: false,
-          },
-        });
-      } else {
-        setModal({
-          isOpen: true,
-          event: 'registration-failed',
-        });
-      }
-    } catch (error) {
-      setModal({
-        isOpen: true,
-        event: 'registration-failed',
-      });
-    }
-  };
-
-  const handleClick = (pln) => setDonate(pln);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setCaptchaSize(window.innerWidth < 1200 ? 'compact' : 'normal');
-    };
-
-    handleResize();
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
 
   useEffect(() => {
     const transactionId = searchParams.get('transactionId');
@@ -140,6 +27,8 @@ const Registry = () => {
       pollPaymentStatus(transactionId);
     }
   }, []);
+
+  const p24Images = Array.from({ length: 25 }, (_, i) => i + 1);
 
   const pollPaymentStatus = async (transactionId) => {
     const maxAttempts = 15;
@@ -188,194 +77,9 @@ const Registry = () => {
           </div>
         </div>
       </div>
-      <section className="Registry__form">
-        <div className="container Registry__form-container">
-          <h2 className="Registry__form-title">
-            {t('REGISTRATION.FORM_TITLE')}
-          </h2>
 
-          <form onSubmit={handleSubmit}>
-            <div className="Registry__form__fieldsGrid">
-              {[
-                {
-                  label: t('REGISTRATION.FORM_FULLNAME'),
-                  name: 'name',
-                  type: 'text',
-                },
-                {
-                  label: t('REGISTRATION.FORM_EMAIL'),
-                  name: 'email',
-                  type: 'email',
-                },
-                {
-                  label: t('REGISTRATION.FORM_STREET'),
-                  name: 'street',
-                  type: 'text',
-                },
-                {
-                  label: t('REGISTRATION.FORM_BUILDING'),
-                  name: 'buildingNumber',
-                  type: 'text',
-                },
-                {
-                  label: t('REGISTRATION.FORM_INDEX'),
-                  name: 'zipCode',
-                  type: 'text',
-                },
-                {
-                  label: t('REGISTRATION.FORM_CITY'),
-                  name: 'city',
-                  type: 'text',
-                },
-              ].map(({ label, name, type }) => (
-                <div className="Registry__form-field" key={name}>
-                  <label className="Registry__form-label">{label}</label>
-                  <input
-                    className="Registry__form-input"
-                    type={type}
-                    name={name}
-                    value={formData[name]}
-                    onChange={handleChange}
-                    required
-                  />
-                </div>
-              ))}
-            </div>
-
-            <div className="Registry__form__checkbox">
-              <Checkbox
-                name="privacyPolicyAndRegulaminAccepted"
-                checked={formData.consents.privacyPolicyAndRegulaminAccepted}
-                onChange={handleConsentChange}
-              />
-              <p className="Registry__form__checkbox-description">
-                {t(`REGISTRATION.FORM_RIGHTS.1.1`)}
-
-                <NavLink
-                  className="Registry__form__checkbox-description 
-                  Registry__form__checkbox-description--link"
-                  to={'/klauzula-informacyjna'}>
-                  {t(`REGISTRATION.FORM_RIGHTS.1.2`)}
-                </NavLink>
-
-                {t(`REGISTRATION.FORM_RIGHTS.1.3`)}
-
-                <NavLink
-                  className="Registry__form__checkbox-description 
-                  Registry__form__checkbox-description--link"
-                  to={'/regulamin'}>
-                  {t(`REGISTRATION.FORM_RIGHTS.1.4`)}
-                </NavLink>
-
-                {t(`REGISTRATION.FORM_RIGHTS.1.5`)}
-              </p>
-            </div>
-
-            <div className="Registry__form__checkbox">
-              <Checkbox
-                name="dataProcessingConsentAccepted"
-                checked={formData.consents.dataProcessingConsentAccepted}
-                onChange={handleConsentChange}
-              />
-
-              <p className="Registry__form__checkbox-description">
-                {t(`REGISTRATION.FORM_RIGHTS.2`)}
-              </p>
-            </div>
-
-            <div className="Registry__form__checkbox">
-              <Checkbox
-                name="marketingConsent"
-                checked={formData.consents.marketingConsent}
-                onChange={handleConsentChange}
-                required={false}
-              />
-              <p className="Registry__form__checkbox-description">
-                {t(`REGISTRATION.FORM_RIGHTS.3`)}
-              </p>
-            </div>
-
-            <p className="Registry__form-description">
-              {t('REGISTRATION.MESSAGE_TEXT')}
-            </p>
-
-            <p className="Registry__form-required">
-              {t('REGISTRATION.MESSAGE_REQUIRED')}
-            </p>
-
-            <HCaptcha
-              className="h-captcha"
-              sitekey={hCaptchaSiteKey} // trzeba założyc konto fundacyjne na hcaptcha.com i pobrac sitekey
-              ref={hcaptchaRef}
-              onVerify={handleHCaptcha}
-              size={captchaSize}
-            />
-
-            <button className="Registry__form-button" type="submit">
-              {t('REGISTRATION.FORMS_BUTTON')}
-            </button>
-          </form>
-        </div>
-      </section>
-
-      <section className="Registry__donate" id="platnosci">
-        <div className="container Registry__donate-container">
-          <h2 className="Registry__donate-title">
-            {' '}
-            {t('REGISTRATION.DONATE_TITLE')}
-          </h2>
-          <p className="Registry__donate-description">
-            {t('REGISTRATION.DONATE_DESCRIPTION')}
-          </p>
-
-          <div className="Registry__donate__buttons">
-            {[100, 200, 250, 400, 500, 1000].map((amount) => (
-              <button
-                key={amount}
-                onClick={() => handleClick(amount)}
-                className={`Registry__donate__buttons-button ${
-                  donate === amount
-                    ? 'Registry__donate__buttons-button--active'
-                    : ''
-                }`}>
-                {`${amount} zł`}
-              </button>
-            ))}
-
-            <div className="Registry__donate__buttons-button Registry__donate__buttons-custom">
-              <input
-                type="number"
-                min="1"
-                placeholder="............"
-                // value={
-                //   ![100, 200, 250, 400, 500].includes(donate) ? donate : ''
-                // }
-                onChange={(e) => handleClick(Number(e.target.value))}
-                className="Registry__donate__buttons-custom-input"
-              />
-              <span className="Registry__donate__buttons-currency">zł</span>
-            </div>
-          </div>
-
-          <div className="Registry__donate__field">
-            <label className="Registry__donate__field-label">
-              {t('REGISTRATION.FORM_EMAIL')}
-            </label>
-            <input
-              className="Registry__donate__field-input"
-              type="email"
-              name="donate-email"
-              value={formData[name]}
-              onChange={handleChange}
-              required
-            />
-          </div>
-
-          <button className="Registry__form-button" type="button">
-            {t('REGISTRATION.DONATE_BUTTON')}
-          </button>
-        </div>
-      </section>
+      <RegistryForm setModal={setModal} />
+      <RegistryDonateForm setModal={setModal} />
 
       <section className="Registry__logos">
         <div className="container">
@@ -410,8 +114,6 @@ const Registry = () => {
         url={modal.url}
         onClose={() => setModal({ isOpen: false })}
       />
-
-      <P24Form p24formIsOpen={p24formIsOpen} />
     </main>
   );
 };
