+++ src/pages/Registry/RegistryForm/registryForm.scss
@@ -1,0 +1,170 @@
+@import '../../../styles/mixins.scss';
+
+.RegistryForm {
+    color: #345b55;
+    margin-bottom: 104px;
+    padding: 0 20px;
+
+    &-container {
+      border: 3px solid #345b55;
+      background-color: #ffffff;
+      border-radius: 20px;
+      padding: 89px 95px 40px;
+
+      @include device(mobile) {
+        padding: 82px 27px 45px;
+      }
+    }
+
+    &-title {
+      //styleName: WP_Title H1';
+      font-family: Poppins;
+      font-size: 48px;
+      font-weight: 700;
+      line-height: 72px;
+      letter-spacing: 0.01em;
+      text-align: center;
+      margin-bottom: 95px;
+
+      @include device(mobile) {
+        font-size: 32px;
+        font-weight: 700;
+        line-height: 48px;
+      }
+    }
+
+    &__fieldsGrid {
+      display: grid;
+      grid-template-rows: 92px auto;
+      grid-template-columns: 1fr 1fr;
+      column-gap: 59px;
+      row-gap: 59px;
+      margin-bottom: 71px;
+
+      @include device(mobile) {
+        grid-template-columns: 1fr;
+      }
+    }
+
+    &-label {
+      //styleName: WP_body copy;
+      font-family: Poppins;
+      font-size: 16px;
+      font-weight: 400;
+      line-height: 22.4px;
+      text-align: left;
+      color: #345b55;
+    }
+
+    &-input {
+      width: 100%;
+      box-sizing: border-box;
+      height: 62px;
+      padding: 13px 29px 13px 29px;
+      gap: 10px;
+      border-radius: 5px;
+      border: none;
+      border: 1px solid #345b55;
+      font-family: Poppins;
+      font-size: 16px;
+      font-weight: 400;
+      line-height: 22.4px;
+      text-align: left;
+     
+      &--big {
+        min-height: 300px;
+        resize: none;
+      }
+
+      &:focus {
+        outline: none;
+        border: 2px solid #345b55;
+      }
+    }
+
+    &__checkbox {
+      display: flex;
+      align-items: center;
+      gap: 15px;
+      margin-bottom: 44px;
+
+      &-description {
+        color: black;
+        font-family: Poppins;
+        font-size: 14px;
+        font-weight: 400;
+        line-height: 21px;
+        text-align: left;
+
+        &--link {
+          color: inherit;
+          text-decoration: none;
+          //styleName: WP_Link;
+          font-family: Poppins;
+          font-size: 14px;
+          font-weight: 500;
+          line-height: 21px;
+          text-align: left;
+          text-decoration: underline;
+        }
+      }
+    }
+
+    &-description {
+      font-family: Poppins;
+      font-size: 16px;
+      font-weight: 500;
+      line-height: 24px;
+      letter-spacing: -0.011em;
+      text-align: left;
+      color: #345b55;
+      margin-bottom: 29px;
+    }
+
+    &-required {
+      color: black;
+      font-family: Poppins;
+      font-size: 14px;
+      font-weight: 400;
+      line-height: 21px;
+      text-align: left;
+      margin-bottom: 29px;
+    }
+
+    &-button {
+      all: unset;
+      min-width: 246px;
+      padding: 11.5px 49px;
+      gap: 10px;
+      border-radius: 30px;
+      background: #345b55;
+      color: white;
+      font-family: Poppins;
+      font-size: 16px;
+      font-weight: 600;
+      line-height: 24px;
+      letter-spacing: 0.0012em;
+      text-align: center;
+      box-sizing: border-box;
+      transition: 0.2s;
+      display: flex;
+      margin: 0 auto;
+      justify-content: center;
+
+      @include device(mobile) {
+        //styleName: WP_M_Button;
+        font-family: Poppins;
+        font-size: 14px;
+        line-height: 23px;
+        text-align: left;
+        padding: 9px 25px;
+        min-width: 210px;
+        
+      }
+
+      &:hover {
+        scale: 103%;
+        cursor: pointer;
+      }
+    }
+  }
